% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esbg.R
\name{plot.esbg}
\alias{plot.esbg}
\title{Plot an ESBG fit}
\usage{
\method{plot}{esbg}(x, ...)
}
\arguments{
\item{x}{An object returned by \code{\link[=esbg]{esbg()}}.}

\item{...}{Unused.}
}
\value{
A ggplot object.
}
\description{
The plot uses a Figure 11-inspired style: ESBG groups are shown in blue and
orange, and group centroids are shown as lighter triangles.
}
