% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esbg.R
\name{print.esbg}
\alias{print.esbg}
\title{Print an ESBG fit}
\usage{
\method{print}{esbg}(x, ...)
}
\arguments{
\item{x}{An object returned by \code{\link[=esbg]{esbg()}}.}

\item{...}{Unused.}
}
\description{
Print an ESBG fit
}
