% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esbg.R
\name{print.summary.esbg}
\alias{print.summary.esbg}
\title{Print an ESBG summary}
\usage{
\method{print}{summary.esbg}(x, ...)
}
\arguments{
\item{x}{An object returned by \code{\link[=summary.esbg]{summary.esbg()}}.}

\item{...}{Unused.}
}
\description{
Print an ESBG summary
}
