% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esbg.R
\name{summary.esbg}
\alias{summary.esbg}
\title{Summarize an ESBG fit}
\usage{
\method{summary}{esbg}(object, ...)
}
\arguments{
\item{object}{An object returned by \code{\link[=esbg]{esbg()}}.}

\item{...}{Unused.}
}
\description{
Summarize an ESBG fit
}
